requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-ollama>=0.0.1
langchain-anthropic>=0.0.1
python-dotenv>=0.19.0
anthropic>=0.3.0
ollama>=0.1.0
streamlit>=1.32.0
beautifulsoup4>=4.12.0
requests>=2.31.0
